move moveObj into _velocity
